Covid_19/pyproject.toml
[project]
name = "covid-19"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.3",
    "plotly>=6.1.2",
    "pycountry-convert>=0.7.2",
    "streamlit>=1.45.1",
    "streamlit-plotly-events>=0.0.6",
]
